% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_common_grid.R
\name{create_common_grid}
\alias{create_common_grid}
\title{Title}
\usage{
create_common_grid(x, models, x_comp, x_class, grid_length)
}
\arguments{
\item{x}{Predictor name}

\item{models}{List of singlegam models}

\item{x_comp}{Complement of x, i.e., all other predictors}

\item{x_class}{Class of predictor}

\item{grid_length}{grid length}
}
\value{
List with a vector defining the common grid and the separate grids.
}
\description{
Title
}
