% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{getmasd}
\alias{getmasd}
\title{Get maximum of the absolute standard deviations}
\usage{
getmasd(mod, grid, nmc, terms)
}
\arguments{
\item{mod}{Object of class "striprawdata", returned from \code{\link{strip_rawdata}}.}

\item{grid}{Grid of values over which to create a simultaneous confidence interval.}

\item{nmc}{Number of Monte Carlo samples.}

\item{terms}{Character vector of smooth terms.}
}
\value{
A vector of maxima of absolute standard deviations.
}
\description{
This code is based on Dr. Gavin Simpson's blog post https://fromthebottomoftheheap.net/2016/12/15/simultaneous-interval-revisited/.
}
\keyword{internal}
