% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate.R
\name{interpolate}
\alias{interpolate}
\title{Title}
\usage{
interpolate(models, grid_list, x, x_class)
}
\arguments{
\item{models}{List of models}

\item{grid_list}{List of grids}

\item{x}{Character with predictor name}

\item{x_class}{Class of variable}
}
\value{
Dataframe with interpolated values
}
\description{
Title
}
