% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metagam.R
\name{metagam}
\alias{metagam}
\title{Meta analyze nonlinear models}
\usage{
metagam(
  fits,
  grid,
  type = "terms",
  terms = NULL,
  method = "fixed",
  restrict_max = NULL,
  restrict_min = NULL
)
}
\arguments{
\item{fits}{List of GAM or GAMM fits of class metagam.}

\item{grid}{Grid over which to predict.}

\item{type}{defaults to "terms"}

\item{terms}{Character vector of terms, smooth or parametric.}

\item{method}{Method of meta analysis. Defaults to "fixed".}

\item{restrict_max}{char}

\item{restrict_min}{char}
}
\value{
An object of type metagam.
}
\description{
Meta analyze nonlinear models
}
