% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_metagam.R
\name{plot.metagam}
\alias{plot.metagam}
\title{Plot estimated smooth terms}
\usage{
\method{plot}{metagam}(x, ...)
}
\arguments{
\item{x}{Object returned by \code{\link{metagam}}.}

\item{...}{Other arguments to plot.}
}
\value{
A ggplot object plotting a smooth term of interest along an axis. The meta-analytic
fit is shown as a solid black line, and the cohort fits are shown as dashed lines, separated by
color codes.
}
\description{
Plot the meta-analytic estimate of a smooth term along with the separate fits in each cohort.
}
\details{
This function currently only works for meta-analytic estimates of a single smooth term,
alternatively meta-analysis of response or link function.
}
