% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heterogeneity.R
\name{plot_heterogeneity_q}
\alias{plot_heterogeneity_q}
\title{Heterogeneity Q-plot}
\usage{
plot_heterogeneity_q(data, xvar, alpha_thresh)
}
\arguments{
\item{data}{data made by \code{\link{make_heterogeneity_data}}.}

\item{alpha_thresh}{Significance level. Defaults to \code{.05}.}
}
\value{
ggproto object
}
\description{
Internal function to plot
the heterogeneity Q values
}
\keyword{internal}
