% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_meta.R
\name{prepare_meta}
\alias{prepare_meta}
\title{Prepare a model for meta analysis by removing all original data}
\usage{
prepare_meta(fit, path = NULL)
}
\arguments{
\item{fit}{A model fitted using \code{mgcv::gam} or \code{mgcv::gamm}.}

\item{path}{Optional path in which to save the object.}
}
\value{
An object of class metagam, which contains only the aggregate data from the model fit.
}
\description{
Prepare a model for meta analysis by removing all original data
}
