% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_metagam.R
\name{print.metagam}
\alias{print.metagam}
\title{Print method for metagam objects.}
\usage{
\method{print}{metagam}(x, ...)
}
\arguments{
\item{x}{Object of class metagam.}

\item{...}{Other arguments (not used).}
}
\value{
The function invisibly returns its input argument \code{x}.
}
\description{
Print method for metagam objects.
}
