% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_strip_rawdata.R
\name{print.striprawdata}
\alias{print.striprawdata}
\title{Print method for striprawdata}
\usage{
\method{print}{striprawdata}(x, ...)
}
\arguments{
\item{x}{Object of class \code{striprawdata}.}

\item{...}{Other arguments.}
}
\value{
The function invisibly returns its argument.
}
\description{
Print method for striprawdata
}
