% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlegam.R
\name{singlegam}
\alias{singlegam}
\title{Prepare a single GAM/GAMM fit for meta analysis}
\usage{
singlegam(model, grid, nmc = 100)
}
\arguments{
\item{model}{Fitted gam object.}

\item{grid}{Grid over which to compute posterior.}

\item{nmc}{Number of Monte Carlo samples.}
}
\value{
An object of class \code{singlegam}.
}
\description{
Prepare a single GAM/GAMM fit for meta analysis
}
