% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strip_rawdata.R
\name{strip_rawdata}
\alias{strip_rawdata}
\title{Strip rawdata from a generalized additive model}
\usage{
strip_rawdata(model, path = NULL, save_ranges = FALSE)
}
\arguments{
\item{model}{A model fitted using \code{mgcv::gam} or \code{mgcv::gamm}.}

\item{path}{Optional path in which to save the object as a \code{.rds} file.}

\item{save_ranges}{Logical specifying whether to save the ranges of each
variable used by the model. For numeric variables this amounts to the minimum
and maximum, and for factors all levels are saved.}
}
\value{
Model object with individual participant data removed.
}
\description{
This function removes all individual participant data from a generalized additive
model object, while keeping aggregated quantities. The resulting object can be
shared without exposing individual participant data.
}
\details{
If \code{model} is of class \code{"gamm"}, only the \code{"gam"} part of the
model will be kept.

Thin plate regression splines (\code{bs='tp'} and \code{bs='ts'}) and Duchon splines \code{bs='ds'}
are currently not supported, since for these splines \code{mgcv}
requires the unique values of the explanatory variables for each smooth term for the \code{predict}
method to work. Future updates to this package will fix this.
}
\examples{
library(metagam)
library(mgcv)

## Create 5 datasets
set.seed(1234)
datasets <- lapply(1:5, function(x) gamSim(scale = 5, verbose = FALSE))

## Fit a GAM in each dataset, then use strip_rawdata() to remove
## individual participant data
models <- lapply(datasets, function(dat){
  ## This uses the gam() function from mgcv
  model <- gam(y ~ s(x0, bs = "cr") + s(x1, bs = "cr") + s(x2, bs = "cr"), data = dat)
  ## This uses strip_rawdata() from metagam
  strip_rawdata(model)
})

## Next, we meta-analyze the models.
## It is often most convenient to analyze a single term at a time. We focus on s(x1).
meta_analysis <- metagam(models, terms = "s(x1)")

## We can print some information
summary(meta_analysis)

## We can plot the fit
plot(meta_analysis)
}
