% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_metagam.R
\name{summary.metagam}
\alias{summary.metagam}
\title{Summary method for metagam objects}
\usage{
\method{summary}{metagam}(object, ...)
}
\arguments{
\item{object}{A metagam object as returned by \code{\link{metagam}}.}

\item{...}{other arguments (not used).}
}
\value{
A list of summary information.
}
\description{
Summary method for metagam objects
}
