% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_strip_rawdata.R
\name{summary.striprawdata}
\alias{summary.striprawdata}
\title{Summary method for GAMs stripped for rawdata}
\usage{
\method{summary}{striprawdata}(object, ...)
}
\arguments{
\item{object}{Object returned by \code{\link{strip_rawdata}}.}

\item{...}{Other arguments.}
}
\value{
The function returns its input argument, which is printed to the console.
}
\description{
Summary method for GAMs stripped for rawdata
}
