% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metagam.R
\name{metagam}
\alias{metagam}
\title{Meta analyze GAMs}
\usage{
metagam(
  models,
  grid_length = 40,
  probs = c(0.01, 0.025, 0.05, 0.95, 0.975, 0.99)
)
}
\arguments{
\item{models}{list of models}

\item{grid_length}{lenght of grid}

\item{probs}{posterior quantiles to return}
}
\value{
An object of class \code{metagam}
}
\description{
Meta analyze GAMs
}
