% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metagam.R
\name{metagam}
\alias{metagam}
\title{Meta analyze nonlinear models}
\usage{
metagam(fits, grid, type = "terms", terms = NULL, method = "fixed")
}
\arguments{
\item{fits}{List of GAM or GAMM fits of class metagam.}

\item{grid}{Grid over which to predict.}

\item{type}{defaults to "terms"}

\item{terms}{Character vector of terms, smooth or parametric.}

\item{method}{Method of meta analysis. Defaults to "fixed".}
}
\value{
An object of type metagam.
}
\description{
Meta analyze nonlinear models
}
