% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_metagam.R
\name{plot.metagam}
\alias{plot.metagam}
\title{Plot estimated smooth terms}
\usage{
\method{plot}{metagam}(x, ...)
}
\arguments{
\item{x}{Object returned by \code{\link{metagam}}.}

\item{...}{Other arguments to plot.}
}
\value{
A ggplot object.
}
\description{
Plot the meta-analytic estimate of a smooth term along with the separate fits in each cohort.
}
