% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_metagam.R
\name{plot.metagam}
\alias{plot.metagam}
\title{Plot estimated smooth terms}
\usage{
\method{plot}{metagam}(x, ...)
}
\arguments{
\item{x}{metagam object}

\item{...}{Other arguments to plot.}
}
\description{
Plot estimated smooth terms
}
