% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_metagam.R
\name{plot.metagam}
\alias{plot.metagam}
\title{Generic plotting function with S3 dispatch}
\usage{
\method{plot}{metagam}(x, type = 1, ...)
}
\arguments{
\item{x}{metagam object}

\item{type}{Type of plot}

\item{...}{Other arguments to plot.}
}
\description{
Generic plotting function with S3 dispatch
}
