% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_domination.R
\name{plot_domination}
\alias{plot_domination}
\title{Domination plot}
\usage{
plot_domination(
  x,
  axis,
  term,
  relative = TRUE,
  title = "Dominance Plot",
  width = NULL
)
}
\arguments{
\item{x}{Object returned by \code{metagam}}

\item{axis}{Which variable to plot}

\item{term}{Which term to plot}

\item{relative}{Default to TRUE.}

\item{title}{Title of plot.}

\item{width}{Width of bars. Default to \code{NULL}, which means it is automatically
determined based on the minimum grid spacing.}
}
\description{
Plots the (relative) contribution of the individual GAMs to each data point
on a given axis. It shows whether and how parts of the axis are dominated
by certain individual GAMs.
}
