% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_domination.R
\name{plot_domination}
\alias{plot_domination}
\title{Domination plot}
\usage{
plot_domination(x, axis, relative = TRUE, title = "Domination Plot")
}
\arguments{
\item{x}{Object returned by \code{metagam}}

\item{axis}{Axis}

\item{relative}{Default to TRUE.}

\item{title}{Title of plot.}
}
\description{
Plots the (relative) contribution of the individual GAMs to each data point
on a given axis. It shows whether and how parts of the axis are dominated
by certain individual GAMs.
}
