% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heterogeneity.R
\name{plot_heterogeneity}
\alias{plot_heterogeneity}
\title{Title}
\usage{
plot_heterogeneity(x, axis = NULL, term = NULL, type = "Q", alpha = 0.05, ...)
}
\arguments{
\item{x}{metagam object}

\item{axis}{Which axis}

\item{term}{Which term}

\item{type}{Character specifying which type of plot}

\item{alpha}{Significance level}

\item{...}{Other arguments to plot.}
}
\value{
plot
}
\description{
Title
}
