% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heterogeneity.R
\name{plot_heterogeneity}
\alias{plot_heterogeneity}
\title{Heterogeneity Plot}
\usage{
plot_heterogeneity(
  x,
  axis = x$xvars,
  term = NULL,
  type = "Q",
  alpha_thresh = 0.05
)
}
\arguments{
\item{x}{Object returned by \code{\link{metagam}}.}

\item{axis}{Character specifying which variable to plot. Defaults to \code{NULL}; if \code{x} was
fitted with a single term, the explanatory variable corresponding to this term
is selected.}

\item{term}{Character specifying which smooth term to plot. Defaults to \code{NULL}; if \code{x}
was fitted with a single term, this one is taken.}

\item{type}{Character specifying which type of plot. Either \code{"Q"} for the test statistic
or \code{"p"} for the p-value. Defaults to \code{"Q"}.}

\item{alpha_thresh}{Significance level. Defaults to \code{.05}.}
}
\value{
A ggplot object.
}
\description{
Heterogeneity Plot
}
\details{
This plot visualizes the heterogeneity along the given axis, using Cochran's Q test.
}
\examples{
# See the vignette, either at https://lifebrain.github.io/metagam/articles/heterogeneity.html
# or by typing the following in the console:
# vignette("heterogeneity")


}
