% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heterogeneity.R
\name{plot_heterogeneity_p}
\alias{plot_heterogeneity_p}
\title{Heterogeneity p-plot}
\usage{
plot_heterogeneity_p(data, axis, alpha_thresh)
}
\arguments{
\item{data}{data made by \code{\link{make_heterogeneity_data}}.}

\item{axis}{Character specifying which variable to plot. Defaults to \code{NULL}; if \code{x} was
fitted with a single term, the explanatory variable corresponding to this term
is selected.}

\item{alpha_thresh}{Significance level. Defaults to \code{.05}.}
}
\value{
ggproto object
}
\description{
Internal function to plot
the heterogeneity p values
}
