% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_metagam.R
\name{print.summary.metagam}
\alias{print.summary.metagam}
\title{Print output from summary of metagam fit.}
\usage{
\method{print}{summary.metagam}(x, digits = 8, ...)
}
\arguments{
\item{x}{A summary.metagam object}

\item{digits}{Number of digits to print for meta-analytic p-values}

\item{...}{Other arguments}
}
\value{
The function invisibly returns its input argument \code{x}.
}
\description{
Print output from summary of metagam fit.
}
