% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlegam.R
\name{singlegam}
\alias{singlegam}
\title{Prepare a single GAM/GAMM fit for meta analysis}
\usage{
singlegam(
  model,
  nmc = 100,
  grid_length = 30,
  freq = FALSE,
  unconditional = FALSE
)
}
\arguments{
\item{model}{Fitted gam object.}

\item{nmc}{Number of Monte Carlo samples.}

\item{grid_length}{length}

\item{freq}{Logical}

\item{unconditional}{Logical}
}
\value{
An object of class \code{singlegam}.
}
\description{
Prepare a single GAM/GAMM fit for meta analysis
}
