% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_metagam.R
\name{summary.metagam}
\alias{summary.metagam}
\title{Summary method for metagam objects}
\usage{
\method{summary}{metagam}(object, ...)
}
\arguments{
\item{object}{A metagam object as returned by \code{\link{metagam}}.}

\item{...}{other arguments (not used).}
}
\value{
A list of class \code{summary.metagam} containing the following information:
\itemize{
\item \code{meta_pvals}: dataframe with p-values from each individual fit. These can be meta-analytically
  combined using the \code{metap} package.
\item \code{terms}: smooth terms that have been meta-analyzed.
\item \code{method}: method used for meta-analysis. See the \code{metafor} package for detailed description.
\item \code{intercept}: logical specifying whether or not the intercept has been included in the meta-analysis.
\item \code{cohorts}: Number of datasets ("cohorts") used in the meta-analysis.
}
}
\description{
Summary method for metagam objects
}
